Formatted char left string
